# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.ConfigTriggersDto:
    properties:
      organizationsTriggerTime:
        type: integer
      runFailedTriggerTime:
        type: integer
      usersTriggerTime:
        type: integer
    type: object
...
info:
  contact: {}
  title: papagaio-api
  version: 0.1.0
paths:
...
  /gettriggersconfig:
    get:
      description: Get trigger timers
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/dto.ConfigTriggersDto'
      security:
      - ApiKeyToken: []
      summary: Return time triggers
      tags:
      - Triggers
...
  /restarttriggers:
    post:
      description: Restart timers
      parameters:
      - description: ?all
        in: query
        name: all
        type: boolean
      - description: ?organizationsynktrigger
        in: query
        name: organizationsynktrigger
        type: boolean
      - description: ?runsFailedDiscoveryTrigger
        in: query
        name: runsFailedDiscoveryTrigger
        type: boolean
      - description: ?usersSynkTrigger
        in: query
        name: usersSynkTrigger
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ok
      security:
      - ApiKeyToken: []
      summary: restart triggers
      tags:
      - Triggers
  /savetriggersconfig:
    post:
      description: Save trigger timers
      parameters:
      - description: Config triggers
        in: body
        name: configTriggersDto
        required: true
        schema:
          $ref: '#/definitions/dto.ConfigTriggersDto'
      produces:
      - application/json
      responses:
        "200":
          description: ok
      security:
      - ApiKeyToken: []
      summary: Save time triggers
      tags:
      - Triggers
  /starttriggers:
    post:
      description: Start timers
      parameters:
      - description: ?all
        in: query
        name: all
        type: boolean
      - description: ?organizationsynktrigger
        in: query
        name: organizationsynktrigger
        type: boolean
      - description: ?runsFailedDiscoveryTrigger
        in: query
        name: runsFailedDiscoveryTrigger
        type: boolean
      - description: ?usersSynkTrigger
        in: query
        name: usersSynkTrigger
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ok
      security:
      - ApiKeyToken: []
      summary: start triggers
      tags:
      - Triggers
  /stoptriggers:
    post:
      description: Stop timers
      parameters:
      - description: ?all
        in: query
        name: all
        type: boolean
      - description: ?organizationsynktrigger
        in: query
        name: organizationsynktrigger
        type: boolean
      - description: ?runsFailedDiscoveryTrigger
        in: query
        name: runsFailedDiscoveryTrigger
        type: boolean
      - description: ?usersSynkTrigger
        in: query
        name: usersSynkTrigger
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: ok
      security:
      - ApiKeyToken: []
      summary: stop triggers
      tags:
      - Triggers
  /triggersstatus:
    get:
      description: Get triggers status
      produces:
      - application/json
      responses:
        "200":
          description: ok
      security:
      - ApiKeyToken: []
      summary: get triggers status
      tags:
      - Triggers
securityDefinitions:
  ApiKeyToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"